Create find-nth-term.c
There is a series,  S, where the next term is the sum of pervious three terms. Given the first three terms of the series,a ,b , and c respectively, you have to output the nth term of the series using recursion.

Input Format

The first line contains a single integer, .

The next line contains 3 space-separated integers, a ,b , and  c.

Constraints
1<=n<=20
1=<a,b,c<=100

Output Format


Print the nth term of the series, S(n).